refactor product put route

+++ routes/api/product-routes.js
@@ -57,47 +57,52 @@ router.post('/', (req, res) => {
     });
 });
 
-// update product
-router.put('/:id', (req, res) => {
-  // update product data
-  Product.update(req.body, {
-    where: {
-      id: req.params.id,
-    },
-  })
-    .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+router.put('/:id', async (req, res) => {
+  // update product data
+  try {
+    const updatedProd = await Product.update(req.body, {
+      where: {
+        id: req.params.id,
+      },
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
-    });
+
+    const matchTags = await ProductTag.findAll({ where: { product_id: req.params.id } });
+    console.log(matchTags)
+
+    const newTags = req.body.tagIds
+      .filter((tag_id) => !matchTags.includes(tag_id))
+      .map((tag_id) => {
+        return { 
+          product_id: req.params.id,
+          tag_id,
+        }
+      });
+      console.log(newTags);
+      
+      const productTagsToRemove =  [];
+      matchTags.forEach(tag => {
+        if (tag.tag_id !== newTags.tag_id) {
+          productTagsToRemove.push(tag.id)
+        }
+      })
+
+      if (!updatedProd) {
+        res.status(404).json( { message: 'No product to update!'} )
+        return;
+      }
+      
+      ProductTag.destroy({ where: { id: productTagsToRemove } }) 
+      ProductTag.bulkCreate(newTags),
+
+      res.status(200).json(updatedProd);
+
+  } catch (err) {
+    res.status(500).json(err);
+  }
 });
+
+
 
 router.delete('/:id', async (req, res) => {
   try {
